tests/wpt/tests/subapps/remove-error.tentative.https.html

<!DOCTYPE html>
<title>Sub Apps: Error cases for remove()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/subapps-helpers.js"></script>

<body></body>

<script>

promise_test(async t => {
  const iframe = document.createElement('iframe');
  document.body.appendChild(iframe);

  const iframeDOMException = iframe.contentWindow.DOMException;

  // Detach the frame.
  iframe.remove();

  // At this point the iframe is detached and unloaded, and its execution
  // context is gone.
  await promise_rejects_dom(t, 'NotFoundError', iframeDOMException, iframe.contentWindow.subApps.remove(['/sub-app-id']));
}, "The object is no longer associated to a document.");

promise_test(async t => {
  const iframe = document.createElement('iframe');
  document.body.appendChild(iframe);

  const iframeDOMException = iframe.contentWindow.DOMException;
  t.add_cleanup(() => iframe.remove());

  await promise_rejects_dom(t, 'InvalidStateError', iframeDOMException, iframe.contentWindow.subApps.remove(['/sub-app-id']));
}, "API is only supported in top-level browsing contexts.");

promise_test(async t => {
  const cross_origin_url = 'https://example.com/sub-app-id';

  await promise_rejects_dom(t, 'NotSupportedError', window.subApps.remove([cross_origin_url]));
}, 'API supports only same-origin URLs.');

promise_test(async t => {
  const url_1 = '/sub-app-1';
  const url_2 = '/sub-app-2';
  const url_3 = '/sub-app-3';

  let remove_call_params = [url_1, url_2, url_3];

  let mocked_response = [
    { "manifestIdPath": url_1, "resultType": SubAppsServiceRemoveResultType.kOperationError },
    { "manifestIdPath": url_2, "resultType": SubAppsServiceRemoveResultType.kOperationError },
    { "manifestIdPath": url_3, "resultType": SubAppsServiceRemoveResultType.kOperationError }
  ];

  let expected_results = {
    failedApps: {
      [url_1]: "OperationError",
      [url_2]: "OperationError",
      [url_3]: "OperationError"
    }
  };

  await subapps_remove_expect_success_with_result(t, remove_call_params, mocked_response, expected_results);
}, 'Remove call fails.');

promise_test(async t => {
  const url_1 = '/sub-app-1';
  const url_2 = '/sub-app-2';
  const url_3 = '/sub-app-3';

  let remove_call_params = [url_1, url_2, url_3];

  let mocked_response = [
    { "manifestIdPath": url_1, "resultType": SubAppsServiceRemoveResultType.kSuccess },
    { "manifestIdPath": url_2, "resultType": SubAppsServiceRemoveResultType.kSuccess },
    { "manifestIdPath": url_3, "resultType": SubAppsServiceRemoveResultType.kOperationError }
  ];

  let expected_results = {
    removedApps: [url_1, url_2],
    failedApps: {
      [url_3]: "OperationError"
    }
  };

  await subapps_remove_expect_success_with_result(t, remove_call_params, mocked_response, expected_results);
}, 'Remove call fails with mixed results.');

</script>
